requirements.txt
numpy~=1.22.3
pandas~=1.4.2
matplotlib~=3.5.1
Pillow==9.2.0
pdf2image~=1.16.0
datasets==2.3.2
pdfminer
tika~=1.24
pytesseract~=0.3.8
albumentations~=1.1.0
opencv-python==4.6.0.66
tqdm~=4.64.0
sklearn~=0.0
scikit-learn==1.1.1
torchvision~=0.12.0
transformers~=4.26.0
texthero~=1.1.0
stanza~=1.4.0
requests==2.28.1
PyPDF2==2.4.1
boto3==1.24.22
botocore==1.27.22
PyYAML~=6.0
nltk~=3.7
scipy~=1.8.1
scikit-image~=0.19.3
keras~=2.9.0
tensorflow~=2.9.1
imutils~=0.5.4
pyspellchecker~=0.6.3
pymupdf~=1.20.1
openpyxl~=3.0.10
torch~=1.11.0+cu113
ghostscript~=0.7
chardet~=5.1.0
pandarallel~=1.6.4
Levenshtein~=0.20.9
Unidecode~=1.3.6
lbl2vec~=1.0.2